fixed signup bug + added more front-end checks
