+++ pecoranera/src/main/webapp/account.jsp
@@ -19,7 +19,7 @@
     <script src="https://code.jquery.com/jquery-3.7.0.js" integrity="sha256-JlqSTELeR4TLqP0OG9dxM7yDPqX1ox/HfgiSLBj8+kM=" crossorigin="anonymous"></script>
     <script src="https://code.jquery.com/ui/1.13.2/jquery-ui.js" integrity="sha256-xLD7nhI62fcsEZK2/v8LsBcb4lG7dgULkuXoXB/j91c=" crossorigin="anonymous"></script>
     <script src="${pageContext.request.contextPath}/scripts/account.js"></script>
-    
+    <script>var user ="${sessionScope.user}";</script>
 </head>
 <body>
     <jsp:include page="./assets/navbar.jsp">
